LPS-76046 Move BasePortletDataHandlerTestCase to export-import-test-util
