+++ modules/apps/collaboration/blogs/blogs-recent-bloggers-test/src/testIntegration/java/com/liferay/blogs/recent/bloggers/web/internal/exportimport/data/handler/test/RecentBloggersPortletDataHandlerTest.java
@@ -17,10 +17,10 @@ package com.liferay.blogs.recent.bloggers.web.internal.exportimport.data.handler
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.blogs.recent.bloggers.web.constants.RecentBloggersPortletKeys;
 import com.liferay.exportimport.kernel.lar.DataLevel;
+import com.liferay.exportimport.test.util.lar.BasePortletDataHandlerTestCase;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
-import com.liferay.portal.lar.test.BasePortletDataHandlerTestCase;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 
 import org.junit.ClassRule;
