+++ modules/apps/collaboration/comment/comment-test/build.gradle
@@ -9,4 +9,5 @@ dependencies {
 	testIntegrationCompile project(":apps:collaboration:blogs:blogs-api")
 	testIntegrationCompile project(":apps:collaboration:blogs:blogs-test-util")
 	testIntegrationCompile project(":apps:collaboration:comment:comment-page-comments-web")
+	testIntegrationCompile project(":apps:web-experience:export-import:export-import-test-util")
 }
