+++ modules/apps/collaboration/comment/comment-test/src/testIntegration/java/com/liferay/comment/web/internal/exportimport/data/handler/test/PageCommentsPortletDataHandlerTest.java
@@ -19,10 +19,10 @@ import com.liferay.comment.page.comments.web.internal.constants.PageCommentsPort
 import com.liferay.exportimport.kernel.lar.DataLevel;
 import com.liferay.exportimport.kernel.lar.ManifestSummary;
 import com.liferay.exportimport.kernel.lar.StagedModelType;
+import com.liferay.exportimport.test.util.lar.BasePortletDataHandlerTestCase;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
-import com.liferay.portal.lar.test.BasePortletDataHandlerTestCase;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 
 import org.junit.Assert;
