+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-test/build.gradle
@@ -24,4 +24,5 @@ dependencies {
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.portal.upgrade", version: "2.0.0"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.0.0"
 	testIntegrationCompile project(":apps:forms-and-workflow:dynamic-data-lists:dynamic-data-lists-form-web")
+	testIntegrationCompile project(":apps:web-experience:export-import:export-import-test-util")
 }
