+++ modules/apps/foundation/user-groups-admin/user-groups-admin-test/src/testIntegration/java/com/liferay/user/groups/admin/lar/test/UserGroupsAdminPortletDataHandlerTest.java
@@ -16,10 +16,10 @@ package com.liferay.user.groups.admin.lar.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.exportimport.kernel.lar.DataLevel;
+import com.liferay.exportimport.test.util.lar.BasePortletDataHandlerTestCase;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
-import com.liferay.portal.lar.test.BasePortletDataHandlerTestCase;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.user.groups.admin.constants.UserGroupsAdminPortletKeys;
