+++ modules/apps/shopping/shopping-test/build.gradle
@@ -4,4 +4,5 @@ dependencies {
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.shopping.service", version: "2.0.0"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.0.0"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
+	testIntegrationCompile project(":apps:web-experience:export-import:export-import-test-util")
 }
