+++ modules/apps/web-experience/export-import/export-import-test-util/build.gradle
@@ -2,9 +2,10 @@ sourceCompatibility = "1.8"
 targetCompatibility = "1.8"
 
 dependencies {
+	provided group: "com.liferay", name: "com.liferay.registry.api", version: "1.0.0"
 	provided group: "com.liferay", name: "com.liferay.xstream.configurator.api", version: "2.0.0"
 	provided group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.4.0"
-	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.53.0"
+	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	provided group: "com.liferay.portal", name: "com.liferay.portal.test", version: "3.0.0"
 	provided group: "com.liferay.portal", name: "com.liferay.portal.test.integration", version: "3.0.0"
 	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
