+++ modules/apps/web-experience/export-import/export-import-test-util/src/main/java/com/liferay/exportimport/test/util/lar/BasePortletDataHandlerTestCase.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.lar.test;
+package com.liferay.exportimport.test.util.lar;
 
 import com.liferay.exportimport.kernel.lar.DataLevel;
 import com.liferay.exportimport.kernel.lar.ManifestSummary;
@@ -65,6 +65,7 @@ import org.junit.Test;
 /**
  * @author Zsolt Berentey
  * @author Zoltan Csaszi
+ * @author Gergely Mathe
  */
 public abstract class BasePortletDataHandlerTestCase {
 
@@ -109,7 +110,7 @@ public abstract class BasePortletDataHandlerTestCase {
 		Assert.assertEquals(0, getStagedModels().size());
 
 		for (String preferenceKey :
-				portletDataHandler.getDataPortletPreferences()) {
+			portletDataHandler.getDataPortletPreferences()) {
 
 			String portletPreference = portletPreferences.getValue(
 				preferenceKey, StringPool.BLANK);
@@ -523,7 +524,7 @@ public abstract class BasePortletDataHandlerTestCase {
 		checkManifestSummaryReferrerClassNames(manifestSummary);
 
 		for (String manifestSummaryKey :
-				expectedManifestSummary.getManifestSummaryKeys()) {
+			expectedManifestSummary.getManifestSummaryKeys()) {
 
 			String[] keyParts = StringUtil.split(
 				manifestSummaryKey, StringPool.POUND);
@@ -555,7 +556,7 @@ public abstract class BasePortletDataHandlerTestCase {
 		ManifestSummary manifestSummary) {
 
 		for (String manifestSummaryKey :
-				manifestSummary.getManifestSummaryKeys()) {
+			manifestSummary.getManifestSummaryKeys()) {
 
 			Assert.assertFalse(
 				manifestSummaryKey.endsWith(
@@ -760,8 +761,8 @@ public abstract class BasePortletDataHandlerTestCase {
 	protected ZipWriter zipWriter;
 
 	private void _assertControls(
-			PortletDataHandlerControl[] expectedControls,
-			PortletDataHandlerControl[] actualControls)
+		PortletDataHandlerControl[] expectedControls,
+		PortletDataHandlerControl[] actualControls)
 		throws Exception {
 
 		for (PortletDataHandlerControl expectedControl : expectedControls) {
@@ -769,11 +770,11 @@ public abstract class BasePortletDataHandlerTestCase {
 
 			for (PortletDataHandlerControl actualControl : actualControls) {
 				if (expectedControl.getControlName().equals(
-						actualControl.getControlName()) &&
+					actualControl.getControlName()) &&
 					expectedControl.getControlLabel().equals(
 						actualControl.getControlLabel()) &&
 					(expectedControl.isDisabled() ==
-						actualControl.isDisabled())) {
+					 actualControl.isDisabled())) {
 
 					contains = true;
 					break;
