+++ modules/apps/web-experience/portlet-display-template/portlet-display-template-test/build.gradle
@@ -4,4 +4,5 @@ dependencies {
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.dynamic.data.mapping.test.util", version: "2.0.0"
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.portlet.display.template", version: "2.0.0"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
+	testIntegrationCompile project(":apps:web-experience:export-import:export-import-test-util")
 }
